// projects/qa/src/app/models/TestRun.ts
export class TestRun {
  testRunId: number;
  shortDescription: string;
  responsible: string;
  description: string;
  installed_id: number;
  installed_release: number;
  state: string;
  estimatedEffort: number;
  totalTests: number;
  totalTestOk: number;
  totalTestNotOk: number;
  modifier: string;
  creator: string;
  created: string;
  modified: string;

}